feat(v1.4.0): GAMLSS continuous norms for all 7 indicators + fractional age
Replace piecewise-step normative tables with GAMLSS-fitted μ(age)/σ(age)
continuous functions so T-scores vary smoothly with age instead of jumping
at cohort boundaries (6/13/19 yr).

- eeg_analysis.rs: rewrite tbr/apr/faa/paf/rsa/coh/entp_norm as continuous;
  remove COH √T·10 post-transform (standard T=10Z+50 now used uniformly);
  accept age: f64.
- wasm_api.rs: analyze_eeg(samples, age) age parameter f64.
- capability.rs: compute_capability takes f64 age (floor for cohort bucket).
- eegReport.ts: ageFromDob returns day-precision fractional years.
- reportPdf.ts / eegReportHtml.tsx: Math.floor for integer age display.
- docs/algorithm_report.md + CHANGELOG.md: document GAMLSS formulas.

+++ crate/src/wasm_api.rs
@@ -322,7 +322,7 @@ pub fn cmd_stop_acquisition() -> Box<[u8]> {
 ///
 /// `samples_flat`: f32 slice, row-major layout `[sample_idx * 8 + channel_idx]`,
 ///   8 channels, values in µV.
-/// `age`: subject age in years.
+/// `age`: subject age in fractional years (e.g. 12.5 for 12 years 6 months).
 ///
 /// Returns a JSON string:
 /// ```json
@@ -335,7 +335,7 @@ pub fn cmd_stop_acquisition() -> Box<[u8]> {
 /// ```
 /// On error: `{"error":"reason","age":…,"cleanEpochs":…,"totalEpochs":…,"durationSec":…}`
 #[wasm_bindgen]
-pub fn analyze_eeg(samples_flat: &[f32], age: u32) -> String {
+pub fn analyze_eeg(samples_flat: &[f32], age: f64) -> String {
     let result = crate::eeg_analysis::analyze_eeg_internal(samples_flat, age);
     crate::eeg_analysis::result_to_json(&result)
 }
